+++ pkg/project/project.go
@@ -5,7 +5,7 @@ var (
 	gitSHA      = "n/a"
 	name        = "config-controller"
 	source      = "https://github.com/giantswarm/config-controller"
-	version     = "0.2.2-dev"
+	version     = "0.2.2"
 )
 
 // AppControlPlaneVersion is always 0.0.0 for control plane app CRs. These CRs