+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add `giantswarm.io/monitoring: "true"` label to the Service to make sure the
+  app is scraped by the new monitoring platform.
+
 ## [0.2.1] - 2021-01-14
 
 ### Fixed