+++ CHANGELOG.md
@@ -7,17 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2021-01-19
+
 ### Fixed
 
 - Add `giantswarm.io/monitoring: "true"` label to the Service to make sure the
   app is scraped by the new monitoring platform.
+- Resolve catalog URL using storage URL from AppCatalog CR rather than using
+  simple format string.
 
 ## [0.2.1] - 2021-01-14
 
 ### Fixed
 
-- Resolve catalog URL using storage URL from AppCatalog CR rather than using
-  simple format string.
 - Remove old ConfigMap and Secret when a new config version is set.
 
 ### Fixed
@@ -40,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create CLI/daemon scaffolding.
 - Generate application configuration using `generate` command.
 
-[Unreleased]: https://github.com/giantswarm/config-controller/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/giantswarm/config-controller/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/giantswarm/config-controller/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/giantswarm/config-controller/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/giantswarm/config-controller/releases/tag/v0.2.0
 [0.1.0]: https://github.com/giantswarm/config-controller/releases/tag/v0.1.0